Moved nightwatch test into master

+++ Jenkinsfile
@@ -71,10 +71,11 @@ node('vetsgov-general-purpose') {
           'nightwatch-e2e': {
             sh "export IMAGE_TAG=${commonStages.IMAGE_TAG} && docker-compose -p nightwatch up -d && docker-compose -p nightwatch run --rm --entrypoint=npm -e BABEL_ENV=test -e BUILDTYPE=vagovprod vets-website --no-color run nightwatch:docker"
           },
-
-          'nightwatch-accessibility': {
-            sh "export IMAGE_TAG=${commonStages.IMAGE_TAG} && docker-compose -p accessibility up -d && docker-compose -p accessibility run --rm --entrypoint=npm -e BABEL_ENV=test -e BUILDTYPE=vagovprod vets-website --no-color run nightwatch:docker -- --env=accessibility"
-          },
+          if (commonStages.IS_PROD_BRANCH) {
+            'nightwatch-accessibility': {
+              sh "export IMAGE_TAG=${commonStages.IMAGE_TAG} && docker-compose -p accessibility up -d && docker-compose -p accessibility run --rm --entrypoint=npm -e BABEL_ENV=test -e BUILDTYPE=vagovprod vets-website --no-color run nightwatch:docker -- --env=accessibility"
+            },
+          }
 
           cypress: {
             sh "export IMAGE_TAG=${commonStages.IMAGE_TAG} && docker-compose -p cypress up -d && docker-compose -p cypress run --rm --entrypoint=npm -e CI=true -e NO_COLOR=1 vets-website --no-color run cy:test:docker"
@@ -85,7 +86,9 @@ node('vetsgov-general-purpose') {
         throw error
       } finally {
         sh "docker-compose -p nightwatch down --remove-orphans"
-        sh "docker-compose -p accessibility down --remove-orphans"
+        if (commonStages.IS_PROD_BRANCH) {
+          sh "docker-compose -p accessibility down --remove-orphans"
+        }
         sh "docker-compose -p cypress down --remove-orphans"
         step([$class: 'JUnitResultArchiver', testResults: 'logs/nightwatch/**/*.xml'])
       }
